requirements.txt
# File: requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
# NeMo toolkit for ASR (includes PyTorch)
# Pin version for stability, e.g., nemo_toolkit[asr]==1.21.0
# For now, let's use the one from OCR, which implies a recent version.
nemo_toolkit[asr]
onnxruntime>=1.15.0 # Or onnxruntime-gpu if using GPU execution provider
soundfile>=0.12.0
numpy>=1.23.0
python-dotenv>=1.0.0 # For .env file support